Added JSON post in HTTPClient

+++ src/Media/Batch/BatchWatermarker.cpp
@@ -57,8 +57,8 @@ void Media::Batch::BatchWatermarker::ImageOutput(Media::ImageList *imgList, cons
 		Double fontSizePx;
 		UOSInt leng = Text::StrCharCnt(this->watermark);
 		Double sz[2];
-		Int32 iWidth;
-		Int32 iHeight;
+		UInt32 iWidth;
+		UInt32 iHeight;
 		Media::DrawImage *gimg2;
 		Media::DrawBrush *b = tmpImg->NewBrushARGB(0xffffffff);
 		Media::DrawFont *f;
@@ -78,8 +78,8 @@ void Media::Batch::BatchWatermarker::ImageOutput(Media::ImageList *imgList, cons
 			{
 				xRand = Math::Double2Int32(simg->info->dispWidth - sz[0]);
 				yRand = Math::Double2Int32(simg->info->dispHeight - sz[1]);
-				iWidth = Math::Double2Int32(sz[0]);
-				iHeight = Math::Double2Int32(sz[1]);
+				iWidth = (UInt32)Math::Double2Int32(sz[0]);
+				iHeight = (UInt32)Math::Double2Int32(sz[1]);
 				gimg2 = this->deng->CreateImage32(iWidth, iHeight, Media::AT_NO_ALPHA);
 				gimg2->DrawString(0, 0, this->watermark, f, b);
 				gimg2->SetAlphaType(Media::AT_ALPHA);
@@ -88,11 +88,11 @@ void Media::Batch::BatchWatermarker::ImageOutput(Media::ImageList *imgList, cons
 				ImageUtil_ColorReplace32A(bmpBits, iWidth, iHeight, (this->rnd->NextInt30() & 0xffffff) | 0x5f808080);
 				if (revOrder)
 				{
-					this->ablend->Blend(simg->data + Math::Double2Int32(this->rnd->NextDouble() * yRand) * simg->info->storeWidth * 4 + Math::Double2Int32(this->rnd->NextDouble() * xRand) * 4, (OSInt)simg->info->storeWidth << 2, bmpBits + iWidth * 4 * (iHeight - 1), -iWidth * 4, iWidth, iHeight, Media::AT_ALPHA);
+					this->ablend->Blend(simg->data + (UInt32)Math::Double2Int32(this->rnd->NextDouble() * yRand) * simg->info->storeWidth * 4 + Math::Double2Int32(this->rnd->NextDouble() * xRand) * 4, (OSInt)simg->info->storeWidth << 2, bmpBits + iWidth * 4 * (iHeight - 1), -iWidth * 4, iWidth, iHeight, Media::AT_ALPHA);
 				}
 				else
 				{
-					this->ablend->Blend(simg->data + Math::Double2Int32(this->rnd->NextDouble() * yRand) * simg->info->storeWidth * 4 + Math::Double2Int32(this->rnd->NextDouble() * xRand) * 4, (OSInt)simg->info->storeWidth << 2, bmpBits, iWidth * 4, iWidth, iHeight, Media::AT_ALPHA);
+					this->ablend->Blend(simg->data + (UInt32)Math::Double2Int32(this->rnd->NextDouble() * yRand) * simg->info->storeWidth * 4 + Math::Double2Int32(this->rnd->NextDouble() * xRand) * 4, (OSInt)simg->info->storeWidth << 2, bmpBits, iWidth * 4, iWidth, iHeight, Media::AT_ALPHA);
 				}
 				this->deng->DeleteImage(gimg2);
 				tmpImg->DelFont(f);
